+++ cookbooks/printers/CHANGELOG.md
@@ -4,6 +4,7 @@ This file is used to list changes made in each version of printerdrv.
 
 ## 0.5.0:
 
+* Updated with DAC100 specifics.
 * Updated HP Printer Drivers to 2.14
 * Updated Epson Printer Drivers to 2.14
 * Disabled printer sharing